do not try to kill procs if no process group found
git-svn-id: https://tor.ern.nps.edu/svn/proj/labtainer/trunk@1723 ed162e55-3153-4343-a18d-e086fa355ed4

+++ scripts/labtainer-student/lab_sys/sbin/capinout.sh
@@ -19,18 +19,20 @@ END
 trapfun()
 {
     #echo "PROGNAME is $PROGNAME"
-    PGROUP=$(ps x -o "%r %c" | grep $PROGNAME | awk '{print $1}')
-    #echo "PGROUP IS $PGROUP"
-    HAS_ROOT=$(pgrep -u root -g $PGROUP)
-    if [[ -z $HAS_ROOT ]]; then
-        kill -TERM -$PGROUP
-    else
-        sudo kill -TERM -$PGROUP
-    fi
+    PGROUP=$(ps ax -o "%r %c" | grep $PROGNAME | awk '{print $1}')
+    if [[ ! -z $PGROUP ]]; then
+        #echo "PGROUP IS $PGROUP"
+        HAS_ROOT=$(pgrep -u root -g $PGROUP)
+        if [[ -z $HAS_ROOT ]]; then
+            kill -TERM -$PGROUP
+        else
+            sudo kill -TERM -$PGROUP
+        fi
 
-    endtime=`date +"%Y%m%d%H%M%S"`
-    echo "PROGRAM FINISH: $endtime" >> $stdinfile
-    echo "PROGRAM FINISH: $endtime" >> $stdoutfile
+        endtime=`date +"%Y%m%d%H%M%S"`
+        echo "PROGRAM FINISH: $endtime" >> $stdinfile
+        echo "PROGRAM FINISH: $endtime" >> $stdoutfile
+    fi
 }
 tailingquote()
 {
